add-pcf50633-06-RTC_AIE-ioctl
Hey,

the legacy application 'atd' from Russ Nelson/Nils Faerber, used by Qtopia, to
schedule alarms currently fails to start as the the above ioctl is failing.
The other drivers in drivers/rtc implement the above ioctl and we can
implement it too.

The code to mask/unmask the RTC alarm is copied from the set_alarm routine and
adapted to use the reg_set_bit_mask and reg_set_clear_mask.

It is compiling, so it must work. Please welcome me the lkml way ;)


Signed-Off-by: Holger Hans Peter Freyther <zecke@openmoko.org>