fix-jack-debounce.patch
Headphone jack detection is bouncy, it can trigger multiple interrupts
on insertion or removal.  This patch adds a workqueue that waits out the
interrupt spew in 100ms units, and if it sees no more interrupts for 100ms
only then samples and reports the jack state.  I was unable to get a bounce
after 20 or so tries after this.

Signed-off-by: Andy Green <andy@openmoko.com>