fix-force-sdcard-clk-off-when-idle.patch
Existing Glamo bit for stopping SD Card Clock when there is no
transfer taking place does not work.  This patch adds stuff around
the transfer code to force the SD clock up when something is going on
and down when it is idle.  This'll save a little power and noise ;-)

I tested it briefly and was able to SD Boot normally on Sandisk 512M.
Wider testing is appreciated.

Signed-off-by: Andy Green <andy@openmoko.com>