gta02-s3cmci-power-control.patch
This patch gives s3cmci nominal control over power to the WLAN
module.

In face, it only gives it the ability to reset the WLAN module
after module removal/insertion or suspend/resume. Resetting the
module is necessary, because s3cmci requests a full device
initialization from the SD/MMC stack, including CMD5, and this
command is only allowed if the SDIO module has been reset.

Signed-off-by: Werner Almesberger <werner@openmoko.org>