Re: [PATCH] fix root cause of NAND trouble
Ben Dooks wrote:
> My only comment would be
> that a slightly better solution would be to do:

Oh, that's much better, thanks ! I hadn't realized at first that
once could read the FIFO one byte at a time.

The revised patch is below.

- Werner

---------------------------------- cut here -----------------------------------

fix-s3c-nand-read-bytes.patch

With the introduction of optimized OOB reads in nand_read_subpage,
the length of the data requested may not be a multiple of four bytes.

This caused a partial read on the 2440, leading to false ECC errors
and, worse, attempts to "correct" them.

Note that there is a similar issue in s3c2440_nand_write_buf, which
doesn't seem to cause trouble yet.

Signed-off-by: Werner Almesberger <werner@openmoko.org>