debug-s3cmci-fmax.patch
This patch adds the module parameter f_max to s3cmci to limit the
maximum clock frequency of the SD interface.

E.g., when booting with s3cmci.f_max=10000000, the kernel should say
s3c2440-sdi s3c2440-sdi: f_max lowered from 50000000 to 10000000 Hz
and the interface frequency should be limited accordingly.

Setting the limit above the maximum frequency supported by the card
or the interface has no effect, so the option is ignored (and no
"f_max lowered" line is printed).

Setting the limit below the minimum frequency supported by the
interface sets it to that frequency.

Signed-off-by: Werner Almesberger <werner@openmoko.org>