MERGE-via-stable-tracking-hist-ignore-very-short-clicks-s3c24 stable-tracking-hist top was ignore-very-short-clicks-s3c24 / aa3d2f306e4ca5b04a01ca3ad635caae1c1e8e33 ... parent commitmessage: From: Nelson Castillo <nelsoneci@gmail.com> ignore-very-short-clicks-s3c2410_ts.patch
With this patch the issue of the noisy clicks seems to be
alleviated. We ignore the first two points. If a click has less
than 3 points we ignore it.

On a normal click (I mean with at least 3 points) this patch
will delay the first DOWN event by 20 milliseconds.

I'm looking for a better solution but this one works and it should
make the driver more usable.

Signed-off-by: Nelson Castillo <nelsoneci@gmail.com>