Replace "variance" filter with "group" filter
This filter is useful to reject samples that are not reliable. We consider
that a sample is not reliable if it deviates form the Majority.

This filter performs better than the variance filter thus we can replace it.
I will remove the skip filter later.

I haven't seen a bad click in a while. Somebody wake me up.

Signed-off-by: Nelson Castillo <nelsoneci@gmail.com>