Fix s3c_irq_demux_extint8 to handle masked but pending IRQs
Hi,

It was observed that  even after the move to level irq on pcf50633, the
problem of 'the phone doesn't resume if POWER is pressed during
transition to standby' remains.

The original problem was due to a lost edge.

The present problem - When POWER is pressed during the transition,
handle_level_irq runs which looks at the IRQ_DISABLED flag set by
pcf50633_suspend and masks the irq on chip. Now when the cpu wakes up
(level keeps hitting us), s3c_irq_demux_extint8 runs, which tries to
handle all pending irqs. But before it does so, it applies the mask
and hence not all pending irqs were acked and hence, our pmu's irq
(level) caused an irq flood..

This patch clears those interrupts which are masked, but pending. Fixes
the problem for me. Mildly tested.


Signed-off-by: Balaji Rao <balajirrao@openmoko.org>