MERGE-via-stable-tracking-hist-rfkill-support-for-the-ar6000- stable-tracking-hist top was rfkill-support-for-the-ar6000- / e82f8ffa04ebcb05506c8e22268371c3c3b03173 ... parent commitmessage: From: Werner Almesberger <werner@openmoko.org> rfkill support for the AR6000 driver
This patch adds rfkill support to the AR6000 driver. The driver does
not directly implement an rfkill device but uses the help of a special
platform device, such that the latter can retain rfkill state when the
AR6k driver is removed (e.g., because of suspend).

If an attempt is made to bring the driver up (module load, bind, or
resume) while rfkill is blocking, only the driver's data structures
are initialized, but function activation and most of the rest of the
setup is deferred until the rfkill block is removed.

Signed-off-by: Werner Almesberger <werner@openmoko.org>