use msleep instead of mdelay in AR6k
The AR6000 driver uses A_MDELAY with sometimes long delays at several
places. A_MDELAY is just another name for "mdelay", which is a busy
loop. All these delays appear to occur in a context where we can sleep,
so it's better to use "msleep" instead.

This includes the "big fat mdelay(2000)" Balaji found in the suspend
path:
http://lists.openmoko.org/pipermail/openmoko-kernel/2008-December/007171.html

Note that only the ones in DevPollMboxMsgRecv and ar6000_reset_device
are actually invoked in normal use, and only during initialization and
shutdown, respectively. So the impact on overall system performance is
small.

Signed-off-by: Werner Almesberger <werner@openmoko.org>