AR6k: don't retrieve the debug log
The AR6k firmware maintains a log of debugging events that are retrieved
by the host on various occasions, including driver removal. Retrieval of
this data is very slow and seems to have little value.

This patch adds a new configuration option CONFIG_AR6000_WLAN_DEBUG to
enable retrieval of the debugging log. This option is turned off by
default.

Signed-off-by: Werner Almesberger <werner@openmoko.org>