introduce-panic-blink-led-not-using-userspace-omfg.patch
A panic is silent on GTA02, it would be good if we got a little hint
if we are crashing (eg, in suspend / resume) from a panic instead of
a deadlock, etc.  On a normal PC i8042 blinks the keyboard lights if
we panic, this patch causes AUX to flash at 5Hz in event of a panic.

Tested by giving kernel fake root= that didn't exist.

Signed-off-by: Andy Green <andy@openmoko.com>