fix-pcf50633-move-charger-enable-decision-to-pcf50633_usb_curlim_set.patch
At the moment we ended up enabling and disabling the charger
at the time we service USBINS and USBREM events in the ISR.

We did that in machine specific code, but really it is PMU
specific decision, so this patch moves it into PMU code.

We enabled and disabled the charge immediately we saw the
interrupt event from PMU, yet we deferred setting charging
state based on ID and USB enumeration until at least after a
PMU ADC read that was handled in a second interrupt later.

Now the timing is unified to after we determined the charger
type and we enable and disable right there.

Signed-off-by: Andy Green <andy@openmoko.com>