config-add-udf-support.patch
Timo noticed that his compound USB device wasn't getting recognized properly,
it was stuck on the CD-ROM class part of it.  This was accepted by the GTA02
but no scsi device was spawned for it.  He figured out in the end that it is
the lack of a UDF filesystem module that was the problem.

This patch adds UDF and CDROM as a built-in on defconfig-2.6.24, for the
modular configs it just adds it as a module.

Reported-by: Timo Drick <timo@drick.de>
Signed-off-by: Andy Green <andy@openmoko.com>