fix-glamo-suspend-resume-dram-and-engines.patch
Two issues... we never took care to take down engines in suspend
and bring them back in resume.  This was part of the display
corruption that could be seen briefly on resume.  The other issue
that made the "noise" corruption was bad ordering of resume steps.

This patch simplifies (removing needless re-init) resume actions
and makes explicit the suspend and resume steps.  It also adds
code to track which engines are up and push them down in suspend
and bring them back in resume.

The result is no more corruption of display buffer in suspend, it
comes back completely clean.

Signed-off-by: Andy Green <andy@openmoko.com>