deglitch-wlan-power-control.patch
gta02_wlan_power(1) causes a reset each time it's invoked, even if the
device was not turned off.

This causes problems if we use the set_power callback from s3cmci,
because it will try to turn on power even when the device is already
operational.

This patch makes sure we only reset the WLAN if the device was off or
if we don't know the current state.

Doing the "deglitching" in gta02_wlan_power instead of the set_power
handler has the advantage that we're synchronized with any other
agents that may change WLAN power.

Signed-off-by: Werner Almesberger <werner@openmoko.org>