GTA02-specific rfkill framework for WLAN
The patch extends the GTA02 WLAN power management platform device to
register an rfkill handler and to provide a "dock" to which the AR6k
driver can connect to pick up rfkill state and transitions.

We need this indirection because suspend/resume causes removal of the
driver, which therefore cannot retain rfkill state on its own.

Signed-off-by: Werner Almesberger <werner@openmoko.org>