add-force-sdhci-rescan.patch
This provides a way for code to force a mmc rescan by using the s3c hsmmc
platform device as a handle.  It solves the issue of 6410 having two
card detect signals mutually exclusive on one physical ball.

Ben Dooks told he might implement this another way but we need this on GTA03
right now and can migrate it to any official way that turns up.

Signed-off-by: Andy Green <andy@openmoko.com>