Change AR6k driver to module and remove CONFIG_MD
This patch changes the two GTA02 default configurations[1] to build
the AR6k driver and S3C MCI (which is used exclusively by AR6k)
as modules. This way, WLAN can be brought up only if we actually
need it.

The patch also disables CONFIG_MD, which served no clear purpose.

[1] andy removed the moredrivers part despite asking for it
because it can be useful to have it built in without syncing
the rootfs, so the packaging config alone has it as modules.

Signed-off-by: Werner Almesberger <werner@openmoko.org>