make neo1973_pm_gps.c suck less
Here's a bunch of changes to neo1973_pm_gps.c. The thing I set out to
change was just the initial power state, but there's so much wrong in
there, I couldn't resist slaying some evil on the way:

- update the copyright year
- show slightly more confidence in operator precedence
- simplify #ifdef rat's nest a little
- rename "pwron" sysfs file to "power_on", to conform to all the rest
- wrap lines longer than 80 columns
- use PTR_ERR, not cast
- reduce redundancy and increase readability in code patching
  gta01_gps_sysfs_entries
- on GTA02 and if booted by u-boot, GPS is powered on. Power it off,
  like on GTA01 or with Qi.

It's still nasty, but slightly less painful to look at now. Tested on
GTA02 and tested that it still compiles for GTA01.

The name change from "pwron" to "power_on" may be controversial. If
this causes undue misery for distribution makers, please holler.

Signed-off-by: Werner Almesberger <werner@openmoko.org>