introduce-gta02-beachhead.patch
This patch introduces the Openmoko GTA02 machine definition.

Again much of the code is based on Harald Welte's work, although it
has been largely rewritten several times.

This is intended to be the minimum machine definition to boot and
be able to run a rootfs from NAND on GTA02 / Freerunner.  It does
not bring up the framebuffer / Glamo and lacks many other peripheral
drivers from outside the SoC.  But once we have this basis in
mainline kernel, we will be able to introduce the other drivers
and add them here.

Signed-off-by: Andy Green <andy@openmoko.com>