S3C: Add info for supporting circular DMA buffers
The S3C64XX DMA implementation will work a lot better
with the ability to enqueue circular buffers as the
hardware can do it's own linked-list management.

Add a function s3c_dma_has_circular() to show that the
system can do this and a flag for the channel.

Update the s3c24xx/s3c64xx I2S DMA code to deal with
this.

Signed-off-by: Ben Dooks <ben@simtec.co.uk>