fix-gta01-spi-resume-patch
Patch to register the SPI device and thus the save/resume for the display

This patch does *NOT* resolve this problem, but it does make things a bit
better.  The SPI device name changed with 2.6.24, and the change was made for
the gta02 but never propagated to the gta01.  With this change, you'll now see
the entries for the spi bus and the display controller in /sys; and the display
controller suspend/resume functions will be called.

(via BZ 79)

Signed-off-by: Mike Wester <mwester@dis.net>