introduce-charging-led-behaviour.patch
Creates a new behaviour requested by Will that the red LED on GTA02
is lit during battery charging.and goes out when the battery is full.

This is done by leveraging the PMU interrupts, but in one scenario
there is no interrupt that occurs, when the battery is replaced after
being removed with the USB power in all the while.  So a sleepy work
function is started under those circumstances to watch for battery
reinsertion or USB cable pull.

100mA limit was not being observed under some conditions so this was
fixed and tested with a USB cable with D+/D- disconnected.  1A
charger behaviour was also tested.

Showing the charging action exposes some inconsistency in pcf50633
charging action.  If your battery is nearly full, it will keep
charging it at decreasing current even after it thinks it is at
100% capacity for a long while.  But if you pull that same battery
and re-insert it, the charger state machine in pcf50633 believe it is
full and won't charge it.

Signed-off-by: Andy Green <andy@openmoko.com>