fix-pcf50633-kill-white-splash-of-death-on-suspend.patch
mach-gta02 meddles with the regulator platform struct after
it is defined, leading to LCM power getting lost in suspend
despite I set it to be left up.  Fixing this finally removes
the incredibly stubborn white LCM on suspend "flash".

This is also going to be implicated in Sean McNeil's
experience of monochromatic LCM after resume, which was
previously attacked by resetting and re-initing the LCM
from scratch.

In addition, I realized that we take down core_1v3 in
pcf50633 suspend action, this is happening near the
start of suspend, so we are in a meta-race to finish
suspend in a controlled way before the caps on core_1v3
run out (I only saw 23.3uF total).  If it's true, this
is where the weirdo sensitivity to timing during
suspend is coming from.

Therefore in this patch we also remove sleeps and
dev_info() etc (which have to flush on serial console)
from the pc50633 isr workqueue if we are in pcf50633
driver suspend state 1, ie, suspending... because we
don't have time for it.

Signed-off-by: Andy Green <andy@openmoko.com>