workaround-s3c24xx-i2s-stop-live-stream-stall-on-resume.patch Reported-by: Graeme Gregory <graeme@openmoko.org> Basically-solved-by: Graeme Gregory <graeme@openmoko.org>
Graeme found a while back that on resume, IISCON register in
s3c2442 does not show LRCK toggling in LRINDEX bit any more,
causing s3c24xx_snd_lrsync() to timeout and return an error,
aborting restart of any live stream that was playing at
suspend.

I confirmed it was true, meddled around for a bit looking
for some magic to restart LRCK or at least the reporting of
it, and in the end worked around it using the method noted
by Graeme: just ignore LRCK sync if it timed out.  The worst
that could happen would be L and R swap for the duration of
stream that was suspended into but probably not even that.

Signed-off-by: Andy Green <andy@openmoko.com>