ar6k-without-sdio.patch
Make the AR6000 WLAN driver compile after moving it outside the
Atheros SDIO stack. Note that the config option's name changes
as well.

The choice of a non-standard location (drivers/ar6000/) is
intentional. The driver is still very far from being in shape for
mainline inclusion, and the odd location should serve as an
immediate warning.

Not-Yet-Signed-off-by: Werner Almesberger <werner@openmoko.org>