build-introduce-device-specific-named-kernel-binary.patch
This patch changes the ./build script to generate a uImage.bin file with
the build device appended, eg, uImage-GTA02.bin.

The reason for this change is that it is possible to generate rootfs images
now that can target multiple (and probably increasing numbers of) devices
with the one image.  Therefore qi or other bootloader running on the device
is going to have to pick out the right kernel from the rootfs filesystem
for the device it finds it is running on from a set provided down /boot
by the rootfs generator.

/lib/modules for the kernels doesn't get in the way of each other because the
version-specific path is also extended to have the device name as part of it.

2.6.28-GTA02_stable-tracking_c77dda03a7c11666-mokodev

For legacy U-Boot support on GTA01 or GTA02, a symlink to uImage-GTA01.bin or
uImage-GTA02.bin will be needed.

The dfu utility script is updated to take an argument, eg, GTA02

Signed-off-by: Andy Green <andy@openmoko.com>