avoid-double-kfree-ts_filter_destroy_chain.patch
The function ts_filter_mean_destroy is destroying all the
filters of the chain. At the same time each filter is trying
to remove other filters in the chain. Let's fix this by
making each filter ignore it is part of a chain when we are
freeing it.

The bug shows when s3c2410_ts.c is built as a module.

Signed-off-by: Nelson Castillo <nelsoneci@gmail.com>