lis302dl-ack-ff_wu-interrupts.patch
Setup interrupt latching for thresholded operation

From: Simon Kagstrom <simon.kagstrom@gmail.com>

This patch, for stable-tracking, sets up interrupt latching for
thresholded operation, and also acks the interrupt in the handler. This
reduces the interrupt load, and thresholds < 36 are once again allowed
(though not of much use).

The patch also fixes some warnings in the code (including a potential
bug).


Thanks to Werner for pointing this out!

Signed-off-by: Simon Kagstrom <simon.kagstrom@gmail.com>