ar6k-disconnect-more-often.patch
When changing wireless settings, the WMI is often not updated, because
we send a connect but no prior disconnect. The firmware seems to ignore
the connect if already associated.

The disconnect is omitted because the SIOCSIWESSID logic in the
Atheros driver seems to try to prevent multiple disconnects to be sent
to the WMI, which doesn't really seem to be necessary.

This patch makes sure we disconnect when a parameter change has happened.

Signed-off-by: Werner Almesberger <werner@openmoko.org>