ar6k-break-down-insane-allocation.patch
The Atheros WLAN stack kmallocs almost 64kB of contiguous kernel
memory for a structure containing almost entirely buffers.

As is commonly known , this kind of large allocation has a very
high risk of failing as kernel memory fragments during the life of
a system.

This patch allocates the buffers indiviudually, thus shrinking the
structure to a size below 4kB.

Note: this is untested. These buffers are only used with Atheros'
raw interface, which none of the code we have, including wmiconfig,
even seems to know about.

This may fix bug #2133. Code follows Atheros' style, so checkpatch
hates it.

Signed-off-by: Werner Almesberger <werner@openmoko.org>