add-skip_filter-s3c2410_ts.patch
Skip filter for touchscreen values.

Problem: The first and the last sample might be unreliable. We provide
this filter as a separate function in order to keep the event_send_timer_f
function simple. This filter:

* Skips NHEAD points after IE_DOWN
* Skips NTAIL points before IE_UP
* Ignores a click if we have less than (NHEAD + NTAIL + 1) points

Right now the filter is embedded in the driver. We have two #defines with
the parameters SKIP_NHEAD and SKIP_NTAIL. I didn't store their values in
struct skip_filter on purpose.

Signed-off-by: Nelson Castillo <nelsoneci@gmail.com>