Add Linear Filter
This filter allows us to modify the coordinates of the
touchscreen output with the calibration values for the device.

It adds a /sys/ts-calibration directory with 7 files (0..6) that
you can use to change the calibration constants.

This path is not correct. Please do not rely on it because we
will change it really soon.

Signed-off-by: Nelson Castillo <nelsoneci@gmail.com>