AR6k: don't reset on shutdown
The AR6k driver can explicitly reset the module when shutting down the
device, which takes about two seconds and seems to be rather useless,
considering that we disable the entire SDIO function moments later.

This patch adds the configuration option CONFIG_AR6000_WLAN_RESET that
has to be set to maintain the previous default of resetting on shutdown.
The option is turned off by default.

Signed-off-by: Werner Almesberger <werner@openmoko.org>