GTA0[12]: fix calypso sysfs power_on handling
This implements a proper scheme of controlling powering on of the Calypso
GSM modem which should be activated with a "push-button" connected to
MODEM_ON line.

Also a workaround for firmware bug of moko10 and earlier implemented to
force calypso to pull IRQ line down after powering on.

The "reset" sysfs node is left for compatibility with userspace (without it
gsm0710muxd breaks) but it does nothing (NC on hardware level on GTA02
anyway).

I tested flashing the GSM firmware (on GTA02) with this patch (using only
fluid and power_on node) and can confirm that it's working as reliably as
with accessing gpio directly per instructions.

The credit goes to Joerg Reisenweber for discussions and clarifications on
how this should be done properly.

Signed-off-by: Paul Fertser <fercerpav@gmail.com>