PCF50633 core depends on I2C
We select the PCF50633 core part, which depends on I2C, in our Kconfig. However,
make oldconfig does not automatically resolve this "depends on I2C" dependency.
This is known as one can read in the Kconfig documentation. Nevertheless, we
should select I2C for GTA02 as it's needed anyway.

Signed-off-by: Sven Rebhan <odinshorse@googlemail.com>