fix-usb-gadget-udc-not-provoking-curlim-action.patch
When we are informed by UDC stuff that we have been enumerated as
a gadget, we recorded the new current limit but then did nothing.

It meant that on boot, the enumeration happened but we remained
stuck at 100mA current limit.  This patch provokes a call to the
work queue to change the current limit and means that coldplugged
USB to a PC now recognizes it can pull 500mA when enumerated.

Signed-off-by: Andy Green <andy@openmoko.com>