fix-gta03-wlan-pm-and-mmc-rescan.patch
This fixes problems with not enough signals controlled by WLAN enable
and disable (it really does cycle the power) and adapts starting state
for Qi change that defaults it all to off.  It saves about 50mW pulling
the power from it.

It also provokes the SDIO stack to go look for the card after power up
or down so it tracks presence by this power operation too.

Signed-off-by: Andy Green <andy@openmoko.com>