build-discrete-object-dirs.patch
It's a hell of a lot faster swapping between build contexts if we use the O=
option on the kernel makefile to maintain subdirs for each build context
separately.  Note this means you will have to make mrproper the toplevel
kernel build directory, and use eg, ./build GTA02 from now on.

Signed-off-by: Andy Green <andy@openmoko.com>