RFC: don't aggravate kernel panics
When in a kernel panic, we try to turn on the backlight. Because
of the panic, preempt is disabled at that time. Unfortunately,
operating the backlight requires access to I2C, and I2C likes to
sleep.

All this results in the kernel unleashing a nearly endless flood
of complaints complete with stack traces, etc., which often
obscures the real issue.

This patch is one way to improve the situation: it just skips
touching the backlight if we're in an atomic context. However, it
isn't perfect: it would be desirable if the backlight did get
turned on also on a panic. To do so, we would need to use I2C in
a way that doesn't sleep.

So I wonder if this is good enough or if someone has any plans to
make a better solution ?

Signed-off-by: Werner Almesberger <werner@openmoko.org>