Send pen-up events faster (side effect: improve illume keyboard responsiveness)
We were waiting 60ms before reporting a pen-up event to avoid
jitter. Now we wait 8ms (actually 5 with HZ == 200).

Thanks to Marco Trevisan for testing and pointing out that there was a
problem that could be spotted with the illume keyboard.
Note that I used the Terminal mode of the keyboard (no dictionary)
for tests.

I also used touch_test.py and the jitter doesn't seem to be an
issue when drawing lines with the finger.

Reported-by: Marco Trevisan (Treviño) <mail@3v1n0.net>
Signed-off-by: Nelson Castillo <arhuaco@freaks-unidos.net>