consistent defconfig naming ?
The naming of our defconfigs is currently inconsistent: three of them
end in -defconfig and use - also as an internal separator. One sends
in _defconfig but uses - as internal separator, and one ends in
_defconfig without needing an internal separator so far.

All the other defconfigs end in _defconfig. 21 use _ also as
internal separator and 5 use -.

So most of our names differ from the style everyone else is following
and there's also inconsistency among our names.

The git patch below renames them to the dominant all _ scheme.

There's always some latent undesirability for such cleanup patches
that may break some habits or scripts. So if there are strong
sentiments against doing this now, I won't push the issue. But in
general, the sooner one gets those little things done, the better.

Signed-off-by: Werner Almesberger <werner@openmoko.org>