fix-lis302dl-get-status-confirm-data-ready.patch
Level interrupts solve the sticky loss of service from accels issue.
But currently, we get two service actions per one interrupt, leading to
information getting read and sent to the input subsystem twice.

This patch makes the ISR confirm with the lis302dl status register
that there is fresh data before accepting it, it works around the
issue and allows use of the other information in the status reg by
another patch.

Signed-off-by: Andy Green <andy@openmoko.com>