Adapt last Touchscreen Changes from GTA02 to GTA01
Pardon me if I am too verbose but this is my first opportunity to send a patch.

I was wishing to try andy-tracking kernels on GTA01 long ago but It
was  FOSDEM and recent threads on the lists asking for GTA01 that made
me do it.

When I participated on the thread
http://n2.nabble.com/empiezo-a-amargarme-con-los-kernels-necesito-consejo-td2347206.html
 in the Spanish mailing list asking what kernel to use with which
distribution, I responded that only FSO based supported 2.6.28 and
upwards. I could not resist to brag  that I tested 2.6.29 on GTA01 but
only with kexec ( it is possible that my wording was more ambiguous on
the thread).

Nelson Asked if I had problems booting it with u-boot. I tried and was
successful but this time I played a Little with the interface and vala
term and noticed that the on screen keyboard was changing the key
pressed.

Using http://svn.openmoko.org/developers/tick/touchtest/touch_test.py
shown that most of the Up and sometimes the previous Move to where
landing on ([290..310],575).

Revising the code shown that the some of the filters and the option to
disable them were missing on mach-gta01.c.

I have tested the kernel with the patch applied with the filters
disabled and also with them enabled. When the filters are active I
still receive  some misplaced up events but much less (the raster
terminal keyboard is usable).


Signed-off-by:Jose Luis Perez Diez <perezdiez@gmail.com>