fix-pm-gsm-disable-serial-drive-when-off.patch
This patch cleans some cruft from neo1973-pm-gsm.c and makes
some changes to the power code for gsm.

First, on probe it enforces GSM is OFF.

Second, when GSM is OFF, it changes the RTS and TXD pins on
UART 0 to be inputs, instead of driving into a dead subsystem
and costing us ~70mA @ 5V constant dissipation (these 3.3V pins
were found at 1.5V due to that).

Third it asserts the GPIO peripheral function when ON, to keep
the same flow for the new interrupt generation workaround code.

Since several people worked on this area before, and maybe the
assertion of GSM OFF is different (it's OFF from bootloader though)
or maybe something else gets broken, I send the patch for comment
before applying it.

Signed-off-by: Andy Green <andy@openmoko.com>