Subject: ohci_s3c2410: Implement PM resume functionality X-Git-Url: http://git.openmoko.org/?p=kernel.git;a=commitdiff_plain;h=522c644d34461fcd65a94cce1fcac9dc0d6cd828
ohci_s3c2410: Implement PM resume functionality

It was observed that that s3c2410_hc's root hub was not autoresumed when the
system was suspended with the root hub itself autosuspended.

As an effect of this, it was observed that the bluetooth module wouldn't get
detected at the hc's roothub ports when powered up after resume. And when
'lsusb' was executed, the roothub got out of autosuspend by force with a
'roothub lost power, resetting' message.

This patch brings the hcd roothub out of resume at global resume thereby
avoiding the issue.

Signed-off-by: Balaji Rao <balajirrao@openmoko.org>