Front buffer kludge
This makes the GEM "info" ioctl produce an object handle which can be used to refer
to the front buffer (i.e. screen pixmap) in Xorg or otherwise.  This is unfortunately
necessary since this memory is allocated in the kernel by the framebuffer driver, and
is not available to DRM until we move to KMS.

Signed-off-by: Thomas White <taw@bitwiz.org.uk>