Fix oops in GEM object allocation
If the allocation of a new GEM object failed due to there being no available VRAM, then
glamo_gem_object_alloc() would bail out, calling drm_gem_object_unreference() in the
process.  glamodrm_gem_free_object() would then proceed to try to give back the memory
which was unsuccessfully allocated.  This obviously doesn't work, and resulted in an oops.
To fix it, check that the VRAM block handle is non-NULL before freeing it.

Signed-off-by: Thomas White <taw@bitwiz.org.uk>