pcf50633-gpio: Add gpiolib support.
This patch replaces the custom gpio interface of the pcf50633 gpio driver with
gpiolib. Although the pcf50633 offers more functionality then what can be
exposed through the gpiolib api this functionality is seldomly used. Furthermore
functions providing the additional functionality can be modeld on top of
gpiolib.
By using gpiolib it is possible to use generic drivers for devices connected to
one of the pcf50633 gpios.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>