ar6000: send userspace an event on disconnection
Without this userspace gets confused and might fail to obtain a DHCP
lease, bring the interface up etc etc.

This place is definetely not the best or most intuitive to add this code
but when added to the real disconnect callback it messes wpa_supplicant
completely because hardware signals disconnection right after it was
asked to associate (and signals association the next moment).

The particular problem this patch solves is lack of DISCONNECTED event
during suspend/resume; because of it "wpa_cli -a" doesn't relaunch its
script and the interface is not brought up automatically after resume.

Signed-off-by: Paul Fertser <fercerpav@gmail.com>