POWER: bq27x00: Add bq27000 support
This patch adds support for the bq27000 battery to the bq27x00 driver.
The bq27000 is similar to the bq27200 except that it uses the HDQ bus instead
of I2C to communicate with the host system.

The driver is implemented as a platform driver. The driver expects to be
provided with a read callback function through its platform data. The read
function is used assumed to do the lowlevel HDQ handling and return the value of
a certain register.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>