ARM: s3c24xx: Set ARCH_NR_GPIOS according to the selected SoC types.
Currently the code in gpiolib.c tries to register GPIO BANKA-BANKM.
ARCH_NR_GPIOS on the other hand is set only to 256, which would be the
equivalent of BANKA-BANKH. Thus the registration of all other banks will fail.

This patch fixes this by setting S3C_GPIO_END according to the selected SoC
types. S3C_GPIO_END is set to the maximum of the number of GPIOs over all
selected SoC types. Thus it is ensured that memory is not wasted if support for
SoCs with higher GPIO numbers is not built-in.  When registering the bank it is
made sure that banks which are outside of that range are not even tried to be
registered.  Otherwise there would be problems with configs where
CONFIG_S3C24XX_GPIO_EXTRA is set to a non zero value.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>