ASoC: codecs: wm8753: Fix DAI mode switching
The wm8753 codec supports switching between different DAI modes.
The current drivers tries to implement this by changing the dai driver at
runtime. But to properly work this would require support from the ASoC core.

The only difference, from a driver point of view, between the different DAI modes
is how to setup the DAI format.

This patch fixes DAI mode switching, by storing the current format for each DAI
in the drivers private struct and when the DAI mode is changed the format get
reprogrammed according to the requirements of the new DAI mode.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>