jbt6k74: Fix display power on sequence
According to the datasheet, first the regulators need to be turned on, then the
reset pin has to be set to high and then we have to wait 100ms before starting
to send commands.
This patch changes the init sequence accordingly.

It has proven very effective when it comes to fixing the WSOD.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>