powerpc/5200: Fix masking of interrupts for 52xx GPT IRQ.
When using the GPT as interrupt controller interrupts were missing.
It turned out the IrqEn bit of the GPT is not a mask bit, but it also
disables interrupt generation. This modification masks interrupt one
level down the cascade. Note that masking one level down the cascade
is only valid here because the GPT as interrupt controller only serves
one IRQ.

Signed-off-by: Henk Stegeman <henk.stegeman@gmail.com>
[grant.likely@secretlab.ca: added description to the .c file as a comment]
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>