ARM: integrator: add Integrator/CP sched_clock support
Integrator/CP has the 24MHz counter which Versatile and later platforms
also have, which we use for sched_clock support.  Allow this counter to
be used when building a kernel targetting Integrator/CP alone.
Integrator/AP does not have this counter, so we must exclude support
for the Integrator family when this is enabled.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>